+++ CHANGELOG.rst
@@ -14,6 +14,8 @@ Added
 
 -  Documentation files and configuration, using Sphinx.
 -  Implementations for ``autocorr2d`` and ``RealFFTConvolve2D``.
+-  Benchmarking tool for ReconstructionAlgorithm
+-  ``n_iter`` parameter for ReconstructionAlgorithm constructor, so don't need to pass to ``apply``.
 
 Changed
 ~~~~~~~